+++ packages/txwrapper-core/src/test-helpers/index.ts
@@ -1,6 +1,5 @@
 export * from './balancesTransfer';
 export * from './constants';
 export * from './getRegistryPolkadot';
-export * from './itHasCorrectBaseTxInfo';
 export * from './knownChainProperties';
 export * from './signWithAlice';
