+++ packages/txwrapper-dev/src/jest-helpers/itHasCorrectBaseTxInfo.ts
