+++ packages/txwrapper-dev/tsconfig.build.json
@@ -1,0 +1,7 @@
+{
+    "extends": "../../tsconfig.build.json",
+    "compilerOptions": {
+        "outDir": "./lib",
+        "rootDir": "./src",
+    }
+}
